% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{wait_for_finish}
\alias{wait_for_finish}
\title{Wait for a process to finish}
\usage{
wait_for_finish(process, target, timeout1 = 2000, timeout2 = 5000)
}
\arguments{
\item{process}{The process. It should not have \code{stdout} or \code{stderr}
pipes, because that can make it freeze.}

\item{timeout1}{Timeout before message.}

\item{timeout2}{Timeout after message.}
}
\description{
With timeout(s), and interaction, if the session is interactive.
}
\details{
First we wait for 2s. If the process is still alive, then we give
it another 5s, but first let the user know that they can interrupt
the process.
}
\keyword{internal}
