% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xopen-package.R
\docType{package}
\name{xopen-package}
\alias{xopen-package}
\title{xopen: Open System Files, 'URLs', Anything}
\description{
Cross platform solution to open files, directories or 'URLs' with their associated programs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/xopen#readme}
  \item \url{https://r-lib.github.io/xopen/}
  \item Report bugs at \url{https://github.com/r-lib/xopen/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Authors:
\itemize{
  \item Fathi Boudra
  \item Rex Dieter
  \item Kevin Krammer
  \item Jeremy White
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
