% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{xopen}
\alias{xopen}
\title{Open a file, directory or URL}
\usage{
xopen(target = NULL, app = NULL, quiet = FALSE, ...)
}
\arguments{
\item{target}{String, the path or URL to open.}

\item{app}{Specify the app to open \code{target} with, and its arguments,
in a character vector. Note that app names are platform dependent.}

\item{quiet}{Whether to echo the command to the screen, before
running it.}

\item{...}{Additional arguments, not used currently.}
}
\description{
Open a file, directory or URL, using the local platforms conventions,
i.e. associated applications, default programs, etc. This is usually
equivalent to double-clicking on the file in the GUI.
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{xopen("test.R")
xopen("https://ps.r-lib.org")
xopen(tempdir())
}\if{html}{\out{</div>}}
}

